Fix changelog and migration guide entries (#192)
* Fix documentation

* Documentation and unneeded trait bounds

* Safety

* Tests

* More comments and tests

* Rules and build script

* Update docs and tests for contacts -> contact

* Add tests for invalid XML parsing in mj_editing and mj_model

* Add tests for various utility functions and data handling in mj_data and mj_visualization

+++ src/error.rs
@@ -278,10 +278,11 @@ pub enum MjEditError {
     ///
     /// `required_size` follows `snprintf`-style semantics: it is the number of bytes MuJoCo would
     /// write, **not** counting the NUL terminator.  To retry successfully, pass a buffer of at
-    /// least `required_size as usize + 1` bytes.
+    /// least `required_size + 1` bytes.
     XmlBufferTooSmall {
         /// Number of bytes MuJoCo would write (excluding the NUL terminator).
-        required_size: i32,
+        /// Pass a buffer of at least `required_size + 1` bytes to retry.
+        required_size: usize,
     },
 }
